unbound-dhcp-leases-bridge: Fix inotify handling
This patch changes that the script will listen to changes to the
directory instead of the file which got complicated when files got
renamed.

It also processes all changes at the same time and tries finding out
what actions have to be performed in order to avoid unnecessary
iterations.

The script is also limited to process any changes only once every five
seconds to keep resource usage in check on busy systems.

Suggested-by: Anthony Heading <ajrh@ajrh.net>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>