Revert "linux-firmware: Compress firmware on disk"
This reverts commit 77e3829dc1ba7d98516c552e5c1b7c87677dee90.

For the time being, shipping this was found to be too difficult, since
we cannot get linux-firmware down to an acceptable size limit.
Compressing the firmware on installations would work, but takes about 4
minutes on an Intel Xenon CPU alone, hence it is an unacceptable
workload to do for IPFire installation running on weaker hardware.

Therefore, we do not proceed with this at the moment.

Signed-off-by: Peter Müller <peter.mueller@ipfire.org>