Drop powertop
This has been discussed briefly in the telephone conference of
September: powertop is considered to be unnecessary, as IPFire is
optimized for performance, thus interfering with possible power
consumption reducing switches. Also, the need for powertop has been
diminished, given that x86 platforms are highly likely not to run on
batteries, and we are phasing out 32-bit ARM, where this could have been
the case.

Therefore, this patch proposes to drop powertop.

Signed-off-by: Peter Müller <peter.mueller@ipfire.org>
Reviewed-by: Michael Tremer <michael.tremer@ipfire.org>