traffic.cgi: Do not use hard-coded red interface name
If QMI is used the dial in mode has to be set as ppp dialin but the
interface name is red. In such a case the old code tried to display
the stats for the ppp0 interface which is wrong.

This patch fixes this issue by calling the handy function to get
the correct interface name for red.

Signed-off-by: Stefan Schantl <stefan.schantl@ipfire.org>