linux-firmware: Do not ship the qca/ directory
It only contains firmware files for some Qualcomm Bluetooth devices, for
which there is no use on IPFire, since we disabled Bluetooth support in
the kernel a long time ago due to security reasons.

To save some space (~ 1.9 MByte), do not ship these files, and delete
them on existing IPFire installations as well.

Signed-off-by: Peter Müller <peter.mueller@ipfire.org>