squid: Update to 6.1
For details see:
http://lists.squid-cache.org/pipermail/squid-users/2023-July/025929.html

"The Squid HTTP Proxy team is very pleased to announce the availability
of the Squid-6.1 release!

This release is we believe, stable enough for general production use.

Support for Squid-5 bug fixes has now officially ceased. Bugs in 5.x
will continue to be fixed, however the fixes will be added to the 6.x
series. All users of Squid-5.x are encouraged to plan for upgrades."

And:
http://www.squid-cache.org/Versions/v6/RELEASENOTES.html

v6 is running since 6.0.1 here in production use without any seen problems...

Signed-off-by: Matthias Fischer <matthias.fischer@ipfire.org>
Reviewed-by: Peter Müller <peter.mueller@ipfire.org>