extrahd: Rewrite the mount script in shell
This is probably a lot easier than calling all sorts of shell commands
from Perl.

The script has also changed that it will try to mount/umount all
configured mountpoints unless a specific mountpoint is being given.

An initscript will be needed to mount everything when the system is
booting up and umount everything on shutdown.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>