make.sh: Bind-mount all loop devices
There seems to be a different way how to create loop devices. On my
Debian system, the first loop device is a block device with major=7 and
minor=0, the second device is major=7 and minor=1, and so on.

On a system running Grml, the second loop device has major=7 and
minor=32, and all following ones are increasing their minor by 32
as well instead of one.

Since I don't have an easy way to detect this, we will simply bind-mount
all available loop devices in to the build environment.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>