monit: Update to 5.34.3
For details see:
https://mmonit.com/monit/changes/

"Fixed: If the ping statement did not explicitly specify an outgoing
address but a previous ping statement did, the same address was
shared by both statements.

Fixed: Monit may crash upon stopping if the ping statement is used
in conjunction with the address option.

Fixed: If a directory is set in the allow option of the set httpd
statement, instead of a file or string, Monit hangs on startup."

Signed-off-by: Matthias Fischer <matthias.fischer@ipfire.org>
Reviewed-by: Adolf Belka <adolf.belka@ipfire.org>
Signed-off-by: Arne Fitzenreiter <arne_f@ipfire.org>