zlib-ng: Migrate to zlib-ng
This is a new and improved version of zlib that merges various fixes and
optimizations from various contributors.

Amonst those are taking advantage of AVX and instruction sets if they
are available.

This patch adds the new API and a compat library that is a drop-in
replacement for the legacy version of zlib.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>
Signed-off-by: Arne Fitzenreiter <arne_f@ipfire.org>