zabbix_agentd: Set passive check agents to 3 by default on new installations.
Zabbix Agent since v7 by default forks 10 instances to listen for and concurrently execute  incoming (passive) checks. This was only 3 in previous versions and should be plenty on an IPFire instance where resources can be scarce.

Users with an existing installation will have to manually add the parameter to their config if they don't want the Zabbix new default of 10 . This will be documented in the wiki.

Signed-off-by: Robin Roevens <robin.roevens@disroot.org>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>