wireguard: Add a custom routing table for peers
This is a dirty hack to make connections to VPN providers actually work.

We mark all WG packets after encryption and use a secondary routing
table to look up any routes to the peers. That way, we can replace the
default route in the main routing table without having to care about the
special routes there.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>