Revert "tcl: Update to version 9.0.1"
This reverts commit d18f8eed367b1ad870a28963775113bec94e6c0d.

This patch had to be reverted because expect no longer builds in the
toolchain stage. There does not seem to be a fix for this.

In LFS, Arch and Gentoo are also still on version 8.6.

Fedora 42 seems to be on 9.0, but I am not 100% confident expect would
actually build there.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>