wireguard: Don't try to delete any interfaces that don't exist
When the WireGuard interfaces are being destroyed, we are using some
globbing to identify the right interfaces. If there are no interfaces
the globbing string does not match anything and is returned itself. To
avoid an error when trying to delete an interface that never existed, we
configure the shell to never expand empty globbings.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>