openvpn: Ignore existing PID files when starting processes
This is all not very organised and tidy. The init process seems to be
too cautious if there is a PID file left but there should not be any
harm in trying to start the same process twice when in doubt because
after all only one can bind to the same port at a time.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>