ovpnmain.cgi: Never write ncp-disable
This was some compatibility code which was supposed to help us with the
transition towards NCP. Since we are now on OpenVPN 2.6, this version no
longer supports the "ncp-disable" switch and we cannot write it to the
configuration any more. There should always be a default value for data
ciphers.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>