openvpn: Return something if the authentication is in an unknown state
Sometimes, the clients seem to be sending a PUSH_REQUEST which is not
receiving a reply from the server. That is because the authenticator
might not respond at all on the socket.

In that case, we should just aim to restart the authentication and hope
that the client will come back with something that we expected.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>