make.sh: Disable the toolchain during normal builds
This is a precaution because I have seen some files from the toolchain
being baked into the initrd. That should not happen and by keeping PATH
clean, there is no chance for any confusion.

In order to have the tools available that we need to build the initial
system, we will now have to pass a flag.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>