squid: Don't forcibly kill any redirectory processes
This is a race which might cause that when squidGuard pre-compiles any
data, it will be killed too. If that happens, squid will keep forking
squidGuard processes which will be unresponsive (because they are trying
to compile they own databases) and the whole system will become
unresponsive at some point.

There should be no need to perform this action.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>